# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "K-Bond",
  "version": "0.1.0"
 },
 "paths": {
...
  "/book.json": {
   "get": {
    "summary": "Book Json",
    "description": "옛 스냅샷 그대로. 화면이 /api/view 로 옮겨 갈 때까지 남긴다.",
    "operationId": "book_json_book_json_get",
    "parameters": [
     {
      "name": "t",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "T"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/feed.json": {
   "get": {
    "summary": "Feed Json",
    "operationId": "feed_json_feed_json_get",
    "parameters": [
     {
      "name": "t",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "T"
      }
     },
     {
      "name": "since",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 0,
       "title": "Since"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 4000,
       "title": "Limit"
      }
     },
     {
      "name": "h",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "H"
      }
     },
     {
      "name": "d",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "D"
      }
     },
     {
      "name": "k",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "K"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FeedPage"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/view": {
   "get": {
    "summary": "Api View",
    "description": "★화면이 그릴 값. 활성 필터·무크로스·최우선·중앙값·히트맵을 여기서 계산한다.\n\nlane: ktb | msb | nhb | cr · ttl: def | half | inf\ncls·rt: 크레딧에서 고른 종별·등급(등급 커브 선택에 쓴다)",
    "operationId": "api_view_api_view_get",
    "parameters": [
     {
      "name": "t",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "T"
      }
     },
     {
      "name": "lane",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "ktb",
       "title": "Lane"
      }
     },
     {
      "name": "ttl",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "def",
       "title": "Ttl"
      }
     },
     {
      "name": "cls",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cls"
      }
     },
     {
      "name": "rt",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Rt"
      }
     },
     {
      "name": "code",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Code"
      }
     },
     {
      "name": "agg",
      "in": "query",
      "required": false,
      "schema": {
       "type": "number",
       "default": 0.0,
       "title": "Agg"
      }
     },
     {
      "name": "T",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "T"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/View"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/events": {
   "get": {
    "summary": "Events",
    "description": "SSE. 책이 바뀔 때만 민다. 15초 조용하면 하트비트(끊김 오인 방지).",
    "operationId": "events_events_get",
    "parameters": [
     {
      "name": "t",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "T"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Basket": {
    "properties": {
     "t": {
      "type": "integer",
      "title": "T"
     },
     "lo": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lo"
     },
     "hi": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Hi"
     },
     "sec": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sec"
     },
     "rt": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rt"
     },
     "a": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "A"
     },
     "asrc": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "stated",
         "implied",
         "bare",
         "default",
         "oddlot",
         "inherit"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Asrc"
     },
     "d": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "D"
     },
     "ms": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ms",
      "description": "이 니즈에 맞는 오퍼 수"
     },
     "bo": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bo",
      "description": "그중 가장 싼 것"
     }
    },
    "type": "object",
    "required": [
     "t"
    ],
    "title": "Basket",
    "description": "범주 콜 = 크레딧의 «매수면». 종목을 안 찍고 (잔존, 섹터, 등급)으로 온다."
   },
   "BondRow": {
    "properties": {
     "c": {
      "type": "string",
      "title": "C"
     },
     "nm": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Nm"
     },
     "full": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Full"
     },
     "alias": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Alias"
     },
     "ten": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ten"
     },
     "mat": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mat"
     },
     "mp": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mp"
     },
     "mpd": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mpd"
     },
     "fa": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Quote"
       },
       {
        "type": "null"
       }
      ],
      "description": "최우선 오퍼"
     },
     "fb": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Quote"
       },
       {
        "type": "null"
       }
      ],
      "description": "최우선 비드"
     },
     "mid": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mid"
     },
     "fill": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fill",
      "description": "당일 마지막 체결 {t, y}"
     },
     "bench": {
      "type": "boolean",
      "title": "Bench",
      "default": false
     },
     "next": {
      "type": "boolean",
      "title": "Next",
      "default": false
     },
     "today": {
      "type": "boolean",
      "title": "Today",
      "default": false
     },
     "n": {
      "type": "integer",
      "title": "N",
      "default": 0
     },
     "nq": {
      "type": "integer",
      "title": "Nq",
      "default": 0
     },
     "npx": {
      "type": "integer",
      "title": "Npx",
      "default": 0
     },
     "hn": {
      "type": "integer",
      "title": "Hn",
      "default": 0
     },
     "pv": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pv",
      "description": "어제 책의 그 종목"
     }
    },
    "type": "object",
    "required": [
     "c"
    ],
    "title": "BondRow",
    "description": "종목 한 줄 — 서버가 활성 필터·무크로스·최우선·mid 를 계산해 준 것."
   },
   "CreditBucket": {
    "properties": {
     "k": {
      "type": "string",
      "title": "K"
     },
     "cls": {
      "type": "string",
      "title": "Cls"
     },
     "rt": {
      "type": "string",
      "title": "Rt"
     },
     "est": {
      "type": "boolean",
      "title": "Est",
      "description": "등급이 문면이 아니라 집계에서 온 것",
      "default": false
     },
     "n": {
      "type": "integer",
      "title": "N"
     },
     "mb": {
      "type": "integer",
      "title": "Mb",
      "default": 0
     },
     "bp_med": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bp Med"
     },
     "ytm_med": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ytm Med"
     },
     "ttm_lo": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ttm Lo"
     },
     "ttm_hi": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ttm Hi"
     }
    },
    "type": "object",
    "required": [
     "k",
     "cls",
     "rt",
     "n"
    ],
    "title": "CreditBucket",
    "description": "종별×등급 버킷. 값은 중앙값이다."
   },
...
   "Curve": {
    "properties": {
     "offers": {
      "items": {
       "$ref": "#/components/schemas/CurveOffer"
      },
      "type": "array",
      "title": "Offers"
     },
     "mp_line": {
      "items": {
       "items": {},
       "type": "array"
      },
      "type": "array",
      "title": "Mp Line",
      "description": "[[잔존, 민평 중앙값, 표본수], …]"
     },
     "mp_n": {
      "type": "integer",
      "title": "Mp N",
      "default": 0
     }
    },
    "type": "object",
    "required": [
     "offers",
     "mp_line"
    ],
    "title": "Curve"
   },
...
   "DealerCards": {
    "properties": {
     "rows": {
      "items": {
       "$ref": "#/components/schemas/DealerCard"
      },
      "type": "array",
      "title": "Rows",
      "default": []
     },
     "n": {
      "type": "integer",
      "title": "N",
      "default": 0
     },
     "n_both": {
      "type": "integer",
      "title": "N Both",
      "default": 0
     },
     "fa_y": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fa Y"
     },
     "fb_y": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fb Y"
     }
    },
    "type": "object",
    "title": "DealerCards"
   },
...
   "FeedPage": {
    "properties": {
     "feed": {
      "items": {
       "$ref": "#/components/schemas/FeedRow"
      },
      "type": "array",
      "title": "Feed"
     },
     "n_seq": {
      "type": "integer",
      "title": "N Seq"
     }
    },
    "type": "object",
    "required": [
     "feed",
     "n_seq"
    ],
    "title": "FeedPage"
   },
   "FeedRow": {
    "properties": {
     "i": {
      "type": "integer",
      "title": "I",
      "description": "적재 일련번호. 화면이 이걸로 이어 붙인다"
     },
     "t": {
      "type": "integer",
      "title": "T"
     },
     "r": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "R",
      "description": "방"
     },
     "k": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "K",
      "description": "종류 QUOTE·AXE·CONFIRM·INQUIRY·THANKS·OTHER"
     },
     "sec": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sec"
     },
     "n": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "N"
     },
     "code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Code"
     },
     "s": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "S",
         "B"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "S"
     },
     "y": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Y"
     },
     "bp": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bp"
     },
     "a": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "A"
     },
     "asrc": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "stated",
         "implied",
         "bare",
         "default",
         "oddlot",
         "inherit"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Asrc"
     },
     "atmp": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Atmp"
     },
     "csrc": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "stated",
         "level",
         "prev"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Csrc"
     },
     "lvl": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "quoted",
         "conv",
         "est"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Lvl"
     },
     "d": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "D"
     },
     "bk": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bk",
      "description": "딜러 키(가림 뒤 라벨)"
     },
     "h": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "H",
      "description": "하우스(가림 뒤 라벨)"
     },
     "raw": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Raw"
     }
    },
    "type": "object",
    "required": [
     "i",
     "t"
    ],
    "title": "FeedRow",
    "description": "메시지 한 줄. 파싱 결과를 원문과 나란히 흘린다."
   },
   "GradeCurve": {
    "properties": {
     "group": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Group"
     },
     "date": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date"
     },
     "label": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Label"
     },
     "pts": {
      "items": {
       "$ref": "#/components/schemas/GradePoint"
      },
      "type": "array",
      "title": "Pts",
      "default": []
     }
    },
    "type": "object",
    "title": "GradeCurve",
    "description": "등급별 민평 커브 한 벌(`sim_portfolio.credit_matrix` 최신 행)."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "Heat": {
    "properties": {
     "cells": {
      "additionalProperties": {
       "$ref": "#/components/schemas/HeatCell"
      },
      "type": "object",
      "title": "Cells"
     },
     "stale": {
      "type": "boolean",
      "title": "Stale",
      "description": "통안 민평이 그날 것이 아니다",
      "default": false
     },
     "rows": {
      "items": {
       "items": {},
       "type": "array"
      },
      "type": "array",
      "title": "Rows"
     },
     "buckets": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Buckets"
     }
    },
    "type": "object",
    "required": [
     "cells",
     "rows",
     "buckets"
    ],
    "title": "Heat"
   },
   "HeatCell": {
    "properties": {
     "med": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Med"
     },
     "n": {
      "type": "integer",
      "title": "N",
      "default": 0
     },
     "lo": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lo"
     },
     "hi": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Hi"
     },
     "est": {
      "type": "integer",
      "title": "Est",
      "description": "그 칸에 든 «추정»(끝전 0.5 가정) 건수",
      "default": 0
     }
    },
    "type": "object",
    "title": "HeatCell"
   },
...
   "Leaderboard": {
    "properties": {
     "rows": {
      "items": {
       "$ref": "#/components/schemas/LeaderRow"
      },
      "type": "array",
      "title": "Rows",
      "default": []
     },
     "n": {
      "type": "integer",
      "title": "N",
      "default": 0
     },
     "shown": {
      "type": "integer",
      "title": "Shown",
      "default": 0
     }
    },
    "type": "object",
    "title": "Leaderboard"
   },
...
   "ObLadder": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "agg": {
      "type": "number",
      "title": "Agg",
      "default": 0.0
     },
     "levels": {
      "items": {
       "$ref": "#/components/schemas/ObLevel"
      },
      "type": "array",
      "title": "Levels",
      "default": []
     },
     "implied": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Implied",
      "description": "교체에서 나온 조건부 호가",
      "default": []
     },
     "basis": {
      "type": "string",
      "enum": [
       "amt",
       "n"
      ],
      "title": "Basis",
      "default": "n"
     },
     "mx": {
      "type": "number",
      "title": "Mx",
      "default": 1
     },
     "sum": {
      "$ref": "#/components/schemas/ObSum",
      "default": {
       "a": 0.0,
       "b": 0.0,
       "na": 0,
       "nb": 0,
       "imp_a": 0.0,
       "imp_b": 0.0
      }
     },
     "best": {
      "$ref": "#/components/schemas/ObBest",
      "default": {}
     }
    },
    "type": "object",
    "required": [
     "code"
    ],
    "title": "ObLadder",
    "description": "종목 하나의 호가 사다리. 막대 폭·나이 칩은 값이 아니라 표현이라 여기 없다."
   },
...
   "PxSeries": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "mp": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mp"
     },
     "lo": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lo"
     },
     "hi": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Hi"
     },
     "t0": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "T0"
     },
     "t1": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "T1"
     },
     "bin": {
      "type": "integer",
      "title": "Bin",
      "default": 600
     },
     "pts": {
      "items": {
       "$ref": "#/components/schemas/PxPoint"
      },
      "type": "array",
      "title": "Pts",
      "default": []
     },
     "act": {
      "items": {
       "$ref": "#/components/schemas/ActBin"
      },
      "type": "array",
      "title": "Act",
      "default": []
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note",
      "description": "표본이 모자라면 그 이유"
     }
    },
    "type": "object",
    "required": [
     "code"
    ],
    "title": "PxSeries",
    "description": "시세 이력. 세로 범위는 규칙이라 서버가 낸다(민평 대칭)."
   },
...
   "SwapBook": {
    "properties": {
     "n": {
      "type": "integer",
      "title": "N",
      "default": 0
     },
     "pairs": {
      "items": {
       "$ref": "#/components/schemas/SwapPair"
      },
      "type": "array",
      "title": "Pairs",
      "default": []
     }
    },
    "type": "object",
    "title": "SwapBook"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "View": {
    "properties": {
     "now": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Now"
     },
     "T": {
      "type": "integer",
      "title": "T"
     },
     "ttl_mode": {
      "type": "string",
      "enum": [
       "def",
       "half",
       "inf"
      ],
      "title": "Ttl Mode",
      "default": "def"
     },
     "ver": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ver"
     },
     "counts": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "Counts"
     },
     "heat": {
      "$ref": "#/components/schemas/Heat"
     },
     "rows": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/BondRow"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rows",
      "description": "lane 이 ktb·msb·nhb 일 때"
     },
     "buckets": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/CreditBucket"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Buckets",
      "description": "lane 이 cr 일 때"
     },
     "curve": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Curve"
       },
       {
        "type": "null"
       }
      ]
     },
     "mtx_group": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mtx Group",
      "description": "credit_matrix 의 bond_type(등급 커브)"
     },
     "ob": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ObLadder"
       },
       {
        "type": "null"
       }
      ],
      "description": "고른 종목의 호가 사다리"
     },
     "dealers": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/DealerCards"
       },
       {
        "type": "null"
       }
      ],
      "description": "누가 어디 서 있나"
     },
     "swap": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/SwapBook"
       },
       {
        "type": "null"
       }
      ],
      "description": "교체 책(국고만)"
     },
     "px": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/PxSeries"
       },
       {
        "type": "null"
       }
      ],
      "description": "고른 종목의 시세 이력"
     },
     "offers": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/CreditQuote"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Offers",
      "description": "lane 이 cr 일 때 고른 버킷의 오퍼"
     },
     "needs": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/Basket"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Needs",
      "description": "lane 이 cr 일 때 매수 니즈"
     },
     "grade_curve": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/GradeCurve"
       },
       {
        "type": "null"
       }
      ],
      "description": "고른 종별·등급의 민평 커브"
     },
     "pulse": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Pulse"
       },
       {
        "type": "null"
       }
      ]
     },
     "events": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/Event"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Events"
     },
     "event_counts": {
      "anyOf": [
       {
        "additionalProperties": {
         "type": "integer"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Event Counts"
     },
     "curve_today": {
      "anyOf": [
       {
        "additionalProperties": {
         "items": {
          "$ref": "#/components/schemas/CurveTodayRow"
         },
         "type": "array"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Curve Today"
     },
     "leaderboard": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Leaderboard"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "T",
     "counts",
     "heat"
    ],
    "title": "View",
    "description": "`/api/view` 의 응답. 화면은 이걸 그대로 그린다."
   }
  }
 }
}